fix: improve dashspend and address piggycards beta feedback (#1433)
* chore: update coil library for kotlin

* fix: include piggycards gift card transactions in the gift card filter

* fix: change terms dialog to ConstraintLayout

* fix: improve handling of out of stock error

* fix: improve sort where each tab has its own sort settings.

* fix: show the gift card specific discount

* fix: add logging to Networks

* fix: coderabbit suggestions

* fix: make some fields optional in GiftCardResponse

* fix: support non whole number denominations

* fix: fix two crashes

* fix: replace Int with Double on getting fixed card discount

* fix: replace Int with Double on getting fixed card discount

* fix: fix issues with All tab and location counts

* fix: in the All tab, don't sort by online first

* fix: update disabled cards list for piggycards

* fix: remember which tab the user is on in SearchFragment

* refactor: simplify FilterOptions

* refactor: setupKeyboardDetection

+++ features/exploredash/src/main/java/org/dash/wallet/features/exploredash/data/explore/model/Merchant.kt
@@ -56,7 +56,7 @@ data class Merchant(
     @Ignore var maxCardPurchase: Double? = null,
     @Ignore var physicalAmount: Int = 0,
     @Ignore var fixedDenomination: Boolean = false,
-    @Ignore var denominations: List<Int> = listOf(),
+    @Ignore var denominations: List<Double> = listOf(),
     @Ignore var giftCardProviders: List<GiftCardProvider> = listOf()
 ) : SearchResult() {
